cleaning

+++ lib/roles/h2/ops-h2.c
@@ -349,7 +349,7 @@ int rops_handle_POLLOUT_h2(struct lws *wsi)
 			return LWS_HP_RET_BAIL_DIE;
 
 		lwsl_notice("%s: scheduling initial window update to %d\n",
-				__func__, wsi->h2.manual_initial_tx_credit);
+				__func__, (int)wsi->h2.manual_initial_tx_credit);
 		pps->u.update_window.credit = wsi->h2.manual_initial_tx_credit;
 		wsi->h2.told_initial = 1;
 		lws_pps_schedule(wsi, pps);
@@ -721,16 +721,16 @@ rops_callback_on_writable_h2(struct lws *wsi)
 	if (wsi->upgraded_to_http2 && !wsi->h2.h2n->pps &&
 	    !lws_h2_tx_cr_get(wsi)) {
 		/*
-		 * other side is not able to cope with us sending DATA
-		 * anything so no matter if we have POLLOUT on our side if it's
+		 * Other side is not able to cope with us sending any DATA,
+		 * so no matter if we have POLLOUT on our side if it's
 		 * DATA we want to send.
 		 *
 		 * Delay waiting for our POLLOUT until peer indicates he has
 		 * space for more using tx window command in http2 layer
 		 */
-		lwsl_notice("%s: %p: skint (%d)\n", __func__, wsi,
-			    wsi->h2.tx_cr);
+		lwsl_notice("%s: %p: skint (%d)\n", __func__, wsi, wsi->h2.tx_cr);
 		wsi->h2.skint = 1;
+
 		return 0;
 	}
